+++ src/controllers/UserController.js
@@ -66,52 +66,57 @@ module.exports = {
                 res.json({err:"Email already in use!"}); 
             }
         }else{
-            res.status(400);
-            res.json({err:"400 Bad Request!"});
+            res.status(422);
+            return res.json({err:"One or more parameters are missing!"});
         }
     },
     async show(req,res){
         let id = req.params.id;
-        const user = await User.findOne({where:{id:id}});
-        if(user!=undefined){
-            let HATEOAS = [
-                {
-                    href: "http://localhost:8080/user/"+user.id,
-                    method: "DELETE",
-                    rel: "delete_user"
-                },
-                {
-                    href: "http://localhost:8080/user",
-                    method: "POST",
-                    rel: "create_new_user"
-                },
-                {
-                    href: "http://localhost:8080/user/"+user.id,
-                    method: "PUT",
-                    rel: "update_user"
-                },
-                {
-                    href: "http://localhost:8080/user/"+user.id,
-                    method: "GET",
-                    rel: "get_user"
-                },
-                {
-                    href: "http://localhost:8080/users",
-                    method: "GET",
-                    rel: "get_all_users"
-                }
-            ]
-            res.status(200);
-            return res.json({user,_links: HATEOAS});
+        if(id){
+            const user = await User.findOne({where:{id:id}});
+            if(user!=undefined){
+                let HATEOAS = [
+                    {
+                        href: "http://localhost:8080/user/"+user.id,
+                        method: "DELETE",
+                        rel: "delete_user"
+                    },
+                    {
+                        href: "http://localhost:8080/user",
+                        method: "POST",
+                        rel: "create_new_user"
+                    },
+                    {
+                        href: "http://localhost:8080/user/"+user.id,
+                        method: "PUT",
+                        rel: "update_user"
+                    },
+                    {
+                        href: "http://localhost:8080/user/"+user.id,
+                        method: "GET",
+                        rel: "get_user"
+                    },
+                    {
+                        href: "http://localhost:8080/users",
+                        method: "GET",
+                        rel: "get_all_users"
+                    }
+                ]
+                res.status(200);
+                return res.json({user,_links: HATEOAS});
+            }else{
+                res.status(404);
+                res.json({err:"404 Not Found!"});
+            }
         }else{
-            res.status(404);
-            res.json({err:"404 Not Found!"});
+            res.status(422);
+            return res.json({err:"One or more parameters are missing!"});
         }
     },
     async update(req,res){
         let {name,cpf,email,password,phone,zipcode,state,city,street,number,complement} = req.body;
         let id = req.params.id;
-        if(name && cpf && email && password && phone && zipcode && state && city && street && number && complement){
+        if(name && cpf && email && password && phone && zipcode && state && city && street && number && complement && id){
             const user = await User.update({name,cpf,email,password,phone,zipcode,state,city,street,number,complement
             },{where:{id:id}});
             let HATEOAS = [
@@ -144,30 +149,45 @@ module.exports = {
             res.status(200);
             return res.json({user,_links:HATEOAS});
         }else{
-            res.status(400);
-            res.json({err:"400 Bad Request!"});
+            res.status(422);
+            return res.json({err:"One or more parameters are missing!"});
         }
     },
     async destroy(req,res){
         let id = req.params.id;
-        const user = await User.destroy({
-            where:{id}
-        });
-        res.status(200);
-        return res.json(user);
+        if(id){
+            const user = await User.destroy({
+                where:{id}
+            });
+            if(user){
+                res.status(200);
+                return res.json(user);
+            }else{
+                res.status(404);
+                return res.json({err:"404 User not found!"});
+            }
+        }else{
+            res.status(422);
+            return res.json({err:"One or more parameters are missing!"});
+        }
     },
     async confirmMail(req,res){
         let uuid = req.params.uuid;
-        let user = await User.findOne({where:{uuid:uuid}});
-        if(user){
-            let verifiedUser = await User.update({
-                isVerified : true
-            },{where:{uuid:uuid}});
-            res.status(200);
-            return res.json(verifiedUser);
+        if(uuid){
+            let user = await User.findOne({where:{uuid:uuid}});
+            if(user){
+                let verifiedUser = await User.update({
+                    isVerified : true
+                },{where:{uuid:uuid}});
+                res.status(200);
+                return res.json(verifiedUser);
+            }else{
+                res.status(404);
+                return res.json({err:"404 User not found!"});
+            }
         }else{
-            res.status(404);
-            return res.json({err:"404 User not found!"});
+            res.status(422);
+            return res.json({err:"One or more parameters are missing!"});
         }
     }
 }
